+++ crates/swc/tests/tsc-references/esnextmodulekindWithES5Target6_es5.2.minified.js
@@ -2,7 +2,7 @@ export function f1() {
     arguments.length > 0 && void 0 !== arguments[0] && arguments[0];
 }
 export function f2() {
-    for(var _len = arguments.length, arg = new Array(_len), _key = 0; _key < _len; _key++)arg[_key] = arguments[_key];
+    for(var _len = arguments.length, arg = Array(_len), _key = 0; _key < _len; _key++)arg[_key] = arguments[_key];
 }
 export default function f3() {
     arguments.length > 0 && void 0 !== arguments[0] && arguments[0];