+++ crates/swc_html_codegen/src/ctx.rs
@@ -2,13 +2,13 @@ use std::ops::{Deref, DerefMut};
 
 use crate::{writer::HtmlWriter, CodeGenerator};
 
-impl<W> CodeGenerator<W>
+impl<'b, W> CodeGenerator<'b, W>
 where
     W: HtmlWriter,
 {
     /// Original context is restored when returned guard is dropped.
     #[inline]
-    pub(super) fn with_ctx(&mut self, ctx: Ctx) -> WithCtx<W> {
+    pub(super) fn with_ctx(&mut self, ctx: Ctx) -> WithCtx<'_, 'b, W> {
         let orig_ctx = self.ctx;
 
         self.ctx = ctx;
@@ -25,25 +25,25 @@ pub(crate) struct Ctx {
     pub need_escape_text: bool,
 }
 
-pub(super) struct WithCtx<'w, I: 'w + HtmlWriter> {
-    inner: &'w mut CodeGenerator<I>,
+pub(super) struct WithCtx<'w, 'a, I: 'w + HtmlWriter> {
+    inner: &'w mut CodeGenerator<'a, I>,
     orig_ctx: Ctx,
 }
 
-impl<'w, I: HtmlWriter> Deref for WithCtx<'w, I> {
-    type Target = CodeGenerator<I>;
+impl<'w, I: HtmlWriter> Deref for WithCtx<'_, 'w, I> {
+    type Target = CodeGenerator<'w, I>;
 
-    fn deref(&self) -> &CodeGenerator<I> {
+    fn deref(&self) -> &CodeGenerator<'w, I> {
         self.inner
     }
 }
-impl<'w, I: HtmlWriter> DerefMut for WithCtx<'w, I> {
-    fn deref_mut(&mut self) -> &mut CodeGenerator<I> {
+impl<'w, I: HtmlWriter> DerefMut for WithCtx<'_, 'w, I> {
+    fn deref_mut(&mut self) -> &mut CodeGenerator<'w, I> {
         self.inner
     }
 }
 
-impl<'w, I: HtmlWriter> Drop for WithCtx<'w, I> {
+impl<'w, I: HtmlWriter> Drop for WithCtx<'_, 'w, I> {
     fn drop(&mut self) {
         self.inner.ctx = self.orig_ctx;
     }